requirements.txt
# Copyright (c) Microsoft Corporation.
# Licensed under the MIT license.

nltk==3.4
pattern
pytorch-pretrained-bert==0.5.1
transformers
tqdm==4.31.1